extract class out into its own file
